Initial project upload
The code is written in 6 parts to keep the code clean.
The following files contain classes made for project: Room.py, Guest.py, Hotel.py, Booking.py
'Main.py' is the main managing part of the code and works as a bridge between user and system.
'Save_Load.py' is written for Saving data as files and Restoring data from files back into the program. 'pandas' library is used here to save data to excel.